Replace Test Runner implementation with enhanced version
- Removed redundant _v2 files after successful replacement
- Enhanced Test Runner now in original test_runner.py location
- Removed fallback logic in _verify_expected_outcome method as requested
- All Phase 15 features preserved: intelligent step interpretation, living document reports, bug reporting, and smart failure handling
- Updated tests to match the enhanced implementation
