+++ tests/test_test_runner.py
@@ -1,42 +1,34 @@
 """
-Tests for the Test Runner Agent.
+Tests for the Enhanced Test Runner Agent.
 """
 
 import json
+from datetime import datetime, timezone
+from pathlib import Path
 from unittest.mock import AsyncMock, Mock, patch
 from uuid import uuid4
 
 import pytest
 
-from datetime import datetime, timezone
-from src.agents.test_runner import (
-    ExecutionMode,
-    TestRunnerAgent,
-    ActionResult,
-    TestStepResult,
-)
+from src.agents.test_runner import TestRunner
 from src.core.types import (
     ActionInstruction,
     ActionType,
-    EvaluationResult,
+    BugReport,
+    BugSeverity,
     GridAction,
     GridCoordinate,
+    StepResult,
+    TestCase,
+    TestCasePriority,
+    TestCaseResult,
     TestPlan,
+    TestReport,
     TestState,
-    TestStep,
     TestStatus,
+    TestStep,
+    TestSummary,
 )
-
-
-@pytest.fixture
-def mock_settings():
-    """Mock settings for testing."""
-    settings = Mock()
-    settings.openai_api_key = "test-key"
-    settings.openai_model = "gpt-4o-mini"
-    settings.openai_temperature = 0.7
-    settings.openai_max_retries = 3
-    return settings
 
 
 @pytest.fixture
@@ -57,661 +49,809 @@ def mock_browser_driver():
 def mock_action_agent():
     """Mock action agent."""
     agent = AsyncMock()
-    # Create a mock GridAction with proper structure
-    mock_coordinate = GridCoordinate(
-        cell="M23",
-        offset_x=0.5,
-        offset_y=0.5,
-        confidence=0.95,
-        refined=False
-    )
-    mock_instruction = ActionInstruction(
-        action_type=ActionType.CLICK,
-        description="Click button",
-        target="button",
-        expected_outcome="Button clicked"
-    )
-    mock_grid_action = GridAction(
-        instruction=mock_instruction,
-        coordinate=mock_coordinate
-    )
-    agent.determine_action = AsyncMock(return_value=mock_grid_action)
     
-    # Mock the new execute_action method
+    # Mock the enhanced action result structure
+    from src.core.enhanced_types import (
+        AIAnalysis,
+        ValidationResult,
+        CoordinateResult,
+        ExecutionResult,
+        EnhancedActionResult
+    )
+    
+    # Track execution order
+    execution_count = [0]
+    
     async def mock_execute_action(test_step, test_context, screenshot=None):
-        return {
-            "action_type": test_step.action_instruction.action_type.value,
-            "validation_passed": True,
-            "validation_reasoning": "Action is valid",
-            "validation_confidence": 0.95,
-            "grid_cell": "M23",
-            "grid_coordinates": (960, 540),
-            "offset_x": 0.5,
-            "offset_y": 0.5,
-            "coordinate_confidence": 0.95,
-            "execution_success": True,
-            "execution_time_ms": 523.4,
-            "screenshot_after": b"mock_screenshot_after",
-            "ai_analysis": {
-                "success": True,
-                "confidence": 0.9,
-                "actual_outcome": "Action completed successfully"
-            }
-        }
+        """Mock execute_action that returns appropriate results based on action type."""
+        # Use execution order to determine the correct outcome
+        # This ensures we return the right outcome regardless of how the test runner
+        # modifies the test step during execution
+        outcomes_by_order = [
+            "Login page is displayed",  # Step 1 - Navigate
+            "Username is entered",  # Step 2 - Type username
+            "Password is entered",  # Step 3 - Type password
+            "User is logged in and redirected to dashboard",  # Step 4 - Click login
+            "User menu dropdown is displayed",  # Step 5 - Click profile (logout case)
+            "User is logged out and redirected to home page"  # Step 6 - Click logout
+        ]
+        
+        idx = execution_count[0]
+        execution_count[0] += 1
+        
+        if idx < len(outcomes_by_order):
+            outcome = outcomes_by_order[idx]
+        else:
+            outcome = "Action completed successfully"
+        
+        return EnhancedActionResult(
+            test_step_id=test_step.step_id,
+            test_step=test_step,
+            test_context=test_context,
+            validation=ValidationResult(
+                valid=True,
+                reasoning="Action is valid and appropriate",
+                confidence=0.95,
+                concerns=[],
+                suggestions=[]
+            ),
+            coordinates=CoordinateResult(
+                grid_cell="M23",
+                grid_coordinates=(960, 540),
+                offset_x=0.5,
+                offset_y=0.5,
+                confidence=0.95,
+                reasoning="Element found successfully",
+                refined=False,
+                refinement_details=None
+            ),
+            execution=ExecutionResult(
+                success=True,
+                execution_time_ms=523.4,
+                error_message=None,
+                error_traceback=None,
+                browser_logs=[],
+                network_activity=[]
+            ),
+            ai_analysis=AIAnalysis(
+                success=True,
+                confidence=0.95,
+                actual_outcome=outcome,
+                matches_expected=True,
+                ui_changes=["Page updated as expected"],
+                recommendations=[],
+                anomalies=[]
+            ),
+            overall_success=True
+        )
     
     agent.execute_action = AsyncMock(side_effect=mock_execute_action)
     return agent
 
 
-# Evaluator agent fixture removed - evaluation now handled by Action Agent's AI analysis
-
-
 @pytest.fixture
-def test_runner_agent(mock_browser_driver, mock_action_agent):
-    """Create a TestRunnerAgent instance for testing."""
-    agent = TestRunnerAgent(
-        browser_driver=mock_browser_driver,
-        action_agent=mock_action_agent
-    )
-    agent._client = AsyncMock()
-    # Ensure the action agent's call_openai is properly mocked if needed
-    if hasattr(mock_action_agent, 'call_openai'):
-        mock_action_agent.call_openai = AsyncMock()
-    return agent
-
-
-@pytest.fixture
-def sample_test_plan():
-    """Create a sample test plan."""
+def sample_hierarchical_test_plan():
+    """Create a sample hierarchical test plan with test cases."""
     plan_id = uuid4()
-    step1_id = uuid4()
-    step2_id = uuid4()
-    step3_id = uuid4()
-    step4_id = uuid4()
     
-    return TestPlan(
-        plan_id=plan_id,
-        name="Login Test",
-        description="Test user login functionality",
-        requirements="Test that users can log in with valid credentials",
-        steps=[
-            TestStep(
-                step_id=step1_id,
-                step_number=1,
-                description="Navigate to login page",
-                action_instruction=ActionInstruction(
-                    action_type=ActionType.NAVIGATE,
-                    description="Navigate to the login page",
-                    target="login page URL",
-                    expected_outcome="Login page displayed"
-                ),
-                dependencies=[],
-                optional=False
-            ),
-            TestStep(
-                step_id=step2_id,
-                step_number=2,
+    # Create test steps for login test case
+    login_steps = [
+        TestStep(
+            step_number=1,
+            description="Navigate to login page",
+            action="Navigate to https://example.com/login",
+            expected_result="Login page is displayed",
+            action_instruction=ActionInstruction(
+                action_type=ActionType.NAVIGATE,
+                description="Navigate to the login page",
+                target="https://example.com/login",
+                expected_outcome="Login page displayed"
+            )
+        ),
+        TestStep(
+            step_number=2,
+            description="Enter username",
+            action="Type 'testuser' in the username field",
+            expected_result="Username is entered",
+            action_instruction=ActionInstruction(
+                action_type=ActionType.TYPE,
                 description="Enter username",
-                action_instruction=ActionInstruction(
-                    action_type=ActionType.TYPE,
-                    description="Enter username in the username field",
-                    target="username field",
-                    value="testuser",
-                    expected_outcome="Username entered"
-                ),
-                dependencies=[step1_id],
-                optional=False
-            ),
-            TestStep(
-                step_id=step3_id,
-                step_number=3,
+                target="username field",
+                value="testuser",
+                expected_outcome="Username entered"
+            )
+        ),
+        TestStep(
+            step_number=3,
+            description="Enter password",
+            action="Type password in the password field",
+            expected_result="Password is entered",
+            action_instruction=ActionInstruction(
+                action_type=ActionType.TYPE,
                 description="Enter password",
-                action_instruction=ActionInstruction(
-                    action_type=ActionType.TYPE,
-                    description="Enter password in the password field",
-                    target="password field",
-                    value="testpass",
-                    expected_outcome="Password entered"
-                ),
-                dependencies=[step1_id],
-                optional=False
-            ),
-            TestStep(
-                step_id=step4_id,
-                step_number=4,
+                target="password field",
+                value="testpass",
+                expected_outcome="Password entered"
+            )
+        ),
+        TestStep(
+            step_number=4,
+            description="Submit login",
+            action="Click the login button",
+            expected_result="User is logged in and redirected to dashboard",
+            action_instruction=ActionInstruction(
+                action_type=ActionType.CLICK,
                 description="Click login button",
-                action_instruction=ActionInstruction(
-                    action_type=ActionType.CLICK,
-                    description="Click the login button",
-                    target="login button",
-                    expected_outcome="User logged in successfully"
-                ),
-                dependencies=[step2_id, step3_id],
-                optional=False
-            ),
-        ],
-        tags=["login", "authentication"]
+                target="login button",
+                expected_outcome="Login successful"
+            )
+        )
+    ]
+    
+    # Create test steps for logout test case
+    logout_steps = [
+        TestStep(
+            step_number=1,
+            description="Open user menu",
+            action="Click on the user profile icon",
+            expected_result="User menu dropdown is displayed",
+            action_instruction=ActionInstruction(
+                action_type=ActionType.CLICK,
+                description="Click user profile icon",
+                target="user profile icon",
+                expected_outcome="User menu opened"
+            )
+        ),
+        TestStep(
+            step_number=2,
+            description="Click logout",
+            action="Click the logout option",
+            expected_result="User is logged out and redirected to home page",
+            action_instruction=ActionInstruction(
+                action_type=ActionType.CLICK,
+                description="Click logout option",
+                target="logout button",
+                expected_outcome="User logged out"
+            )
+        )
+    ]
+    
+    # Create test cases
+    login_case = TestCase(
+        test_id="TC001",
+        name="User Login Test",
+        description="Test user can login with valid credentials",
+        priority=TestCasePriority.HIGH,
+        prerequisites=["Application is accessible", "Test user account exists"],
+        steps=login_steps,
+        postconditions=["User is logged in", "Dashboard is displayed"],
+        tags=["authentication", "login"]
     )
+    
+    logout_case = TestCase(
+        test_id="TC002",
+        name="User Logout Test",
+        description="Test user can logout successfully",
+        priority=TestCasePriority.MEDIUM,
+        prerequisites=["User is logged in"],
+        steps=logout_steps,
+        postconditions=["User is logged out", "Home page is displayed"],
+        tags=["authentication", "logout"]
+    )
+    
+    # Create test plan
+    test_plan = TestPlan(
+        plan_id=plan_id,
+        name="Authentication Flow Test",
+        description="Test user authentication including login and logout",
+        requirements_source="PRD v1.2 - Authentication Requirements",
+        test_cases=[login_case, logout_case],
+        tags=["authentication", "smoke-test"]
+    )
+    
+    return test_plan
 
 
-class TestTestRunnerAgent:
-    """Test cases for TestRunnerAgent."""
+@pytest.fixture
+def test_runner(mock_browser_driver, mock_action_agent, tmp_path):
+    """Create a TestRunner instance for testing."""
+    runner = TestRunner(
+        browser_driver=mock_browser_driver,
+        action_agent=mock_action_agent,
+        report_dir=str(tmp_path / "test_reports")
+    )
+    runner._client = AsyncMock()
+    return runner
+
+
+class TestEnhancedTestRunner:
+    """Test cases for TestRunner."""
     
     @pytest.mark.asyncio
     async def test_execute_test_plan_success(
-        self, test_runner_agent, sample_test_plan
+        self, test_runner, sample_hierarchical_test_plan, tmp_path
     ):
-        """Test successful test plan execution."""
-        # Mock AI response for progress analysis
-        test_runner_agent.call_openai = AsyncMock(return_value={
-            "content": json.dumps({
-                "assessment": "Test progressing well",
-                "concerns": [],
-                "recommendations": ["Continue with remaining steps"]
-            })
-        })
+        """Test successful test plan execution with all test cases passing."""
+        # Mock AI responses
+        test_runner.call_openai = AsyncMock()
+        
+        # Create a list of responses for step interpretations and verifications
+        responses = []
+        
+        # Step interpretations (6 total - 4 for login case, 2 for logout case)
+        responses.append({"content": json.dumps({"actions": [{"type": "navigate", "target": "https://example.com/login", "critical": True, "description": "Navigate to login page"}]})})
+        responses.append({"content": json.dumps({"actions": [{"type": "type", "target": "username field", "value": "testuser", "critical": True, "description": "Type username in the field"}]})})
+        responses.append({"content": json.dumps({"actions": [{"type": "type", "target": "password field", "value": "testpass", "critical": True, "description": "Type password in the field"}]})})
+        responses.append({"content": json.dumps({"actions": [{"type": "click", "target": "login button", "critical": True, "description": "Click the login button"}]})})
+        responses.append({"content": json.dumps({"actions": [{"type": "click", "target": "user profile icon", "critical": True, "description": "Click on the user profile icon"}]})})
+        responses.append({"content": json.dumps({"actions": [{"type": "click", "target": "logout button", "critical": True, "description": "Click the logout option"}]})})
+        
+        # Outcome verifications (6 total - one for each step)
+        responses.append({"content": json.dumps({"success": True, "actual_outcome": "Login page is displayed", "confidence": 0.95, "reason": ""})})
+        responses.append({"content": json.dumps({"success": True, "actual_outcome": "Username is entered", "confidence": 0.95, "reason": ""})})
+        responses.append({"content": json.dumps({"success": True, "actual_outcome": "Password is entered", "confidence": 0.95, "reason": ""})})
+        responses.append({"content": json.dumps({"success": True, "actual_outcome": "User is logged in and redirected to dashboard", "confidence": 0.95, "reason": ""})})
+        responses.append({"content": json.dumps({"success": True, "actual_outcome": "User menu dropdown is displayed", "confidence": 0.95, "reason": ""})})
+        responses.append({"content": json.dumps({"success": True, "actual_outcome": "User is logged out and redirected to home page", "confidence": 0.95, "reason": ""})})
+        
+        # Create a call counter
+        call_count = [0]
+        
+        def mock_call_openai(**kwargs):
+            idx = call_count[0]
+            call_count[0] += 1
+            if idx < len(responses):
+                response = responses[idx]
+                return response
+            print(f"WARNING: Ran out of mock responses at index {idx}")
+            return {"content": json.dumps({"success": False, "error": "No more responses"})}
+        
+        test_runner.call_openai = AsyncMock(side_effect=mock_call_openai)
         
         # Execute test plan
-        result = await test_runner_agent.execute_test_plan(
-            sample_test_plan,
-            initial_url="https://example.com/login"
+        report = await test_runner.execute_test_plan(
+            sample_hierarchical_test_plan,
+            initial_url="https://example.com"
         )
         
-        # Verify
-        assert isinstance(result, TestState)
-        # Debug: print execution history if test failed
-        if result.status != TestStatus.COMPLETED:
-            for i, step_result in enumerate(test_runner_agent._execution_history):
-                print(f"Step {i+1}: success={step_result.success}, result={step_result.actual_result}")
-        assert result.status == TestStatus.COMPLETED
-        assert len(result.completed_steps) == 4
-        assert len(result.failed_steps) == 0
-        assert len(result.skipped_steps) == 0
+        # Verify report structure
+        assert isinstance(report, TestReport)
+        assert report.test_plan_id == sample_hierarchical_test_plan.plan_id
+        assert report.test_plan_name == "Authentication Flow Test"
+        assert report.status == TestStatus.COMPLETED
+        assert len(report.test_cases) == 2
         
-        # Verify browser interactions
-        test_runner_agent.browser_driver.navigate.assert_called_once_with(
-            "https://example.com/login"
-        )
-        # Action Agent now handles screenshots internally, so browser driver screenshot calls are reduced
-        assert test_runner_agent.browser_driver.screenshot.call_count >= 4  # Initial screenshots for each step
-        assert test_runner_agent.action_agent.execute_action.call_count == 4
-        # Evaluation is now handled by Action Agent's AI analysis
+        # Verify test case results
+        login_result = report.test_cases[0]
+        assert login_result.test_id == "TC001"
+        assert login_result.status == TestStatus.COMPLETED
+        assert login_result.steps_completed == 4
+        assert login_result.steps_failed == 0
+        
+        logout_result = report.test_cases[1]
+        assert logout_result.test_id == "TC002"
+        assert logout_result.status == TestStatus.COMPLETED
+        assert logout_result.steps_completed == 2
+        assert logout_result.steps_failed == 0
+        
+        # Verify summary
+        assert report.summary is not None
+        assert report.summary.total_test_cases == 2
+        assert report.summary.completed_test_cases == 2
+        assert report.summary.failed_test_cases == 0
+        assert report.summary.success_rate == 1.0
+        
+        # Verify no bugs reported
+        assert len(report.bugs) == 0
+        
+        # Verify report files were saved
+        report_dir = Path(tmp_path / "test_reports")
+        json_files = list(report_dir.glob("*.json"))
+        md_files = list(report_dir.glob("*.md"))
+        assert len(json_files) > 0
+        assert len(md_files) > 0
     
     @pytest.mark.asyncio
     async def test_execute_test_plan_with_failure(
-        self, test_runner_agent, sample_test_plan, mock_action_agent
+        self, test_runner, sample_hierarchical_test_plan, mock_action_agent
     ):
-        """Test test plan execution with a critical step failure."""
-        # Make step 2 fail via Action Agent's AI analysis
-        async def mock_execute_action_with_failure(test_step, test_context, screenshot=None):
-            if test_step.step_number == 2:
-                return {
-                    "action_type": test_step.action_instruction.action_type.value,
-                    "validation_passed": True,
-                    "grid_cell": "M23",
-                    "grid_coordinates": (960, 540),
-                    "offset_x": 0.5,
-                    "offset_y": 0.5,
-                    "coordinate_confidence": 0.95,
-                    "execution_success": False,
-                    "screenshot_after": b"mock_screenshot_after",
-                    "ai_analysis": {
-                        "success": False,  # Step 2 fails
-                        "confidence": 0.90,
-                        "actual_outcome": "Error: Username field not found",
-                        "anomalies": ["Username field missing"]
-                    }
-                }
-            else:
-                return {
-                    "action_type": test_step.action_instruction.action_type.value,
-                    "validation_passed": True,
-                    "grid_cell": "M23",
-                    "grid_coordinates": (960, 540),
-                    "offset_x": 0.5,
-                    "offset_y": 0.5,
-                    "coordinate_confidence": 0.95,
-                    "execution_success": True,
-                    "screenshot_after": b"mock_screenshot_after",
-                    "ai_analysis": {
-                        "success": True,
-                        "confidence": 0.95,
-                        "actual_outcome": "Action completed successfully"
-                    }
-                }
-        
-        mock_action_agent.execute_action.side_effect = mock_execute_action_with_failure
-        
-        test_runner_agent.call_openai = AsyncMock(return_value={
-            "content": json.dumps({
-                "assessment": "Test failed early",
-                "concerns": ["Critical step failed"],
-                "recommendations": ["Investigate page load issue"]
-            })
-        })
-        
-        # Execute
-        result = await test_runner_agent.execute_test_plan(sample_test_plan)
-        
-        # Debug output
-        print(f"Test execution history:")
-        for i, step_result in enumerate(test_runner_agent._execution_history):
-            print(f"  Step {i+1}: success={step_result.success}, result={step_result.actual_result}")
-        print(f"Completed steps: {len(result.completed_steps)}")
-        print(f"Failed steps: {len(result.failed_steps)}")
-        
-        # Verify
-        assert result.status == TestStatus.FAILED
-        assert len(result.completed_steps) == 1  # Only first step completed
-        assert len(result.failed_steps) == 1  # Second step failed
-    
-    @pytest.mark.asyncio
-    async def test_execute_step_visual_mode(
-        self, test_runner_agent, sample_test_plan
-    ):
-        """Test executing a step in visual mode."""
-        step = sample_test_plan.steps[0]
-        # Set current test plan
-        test_runner_agent._current_test_plan = sample_test_plan
-        
-        # Execute
-        result = await test_runner_agent._execute_step(step, ExecutionMode.VISUAL)
-        
-        # Verify
-        assert isinstance(result, TestStepResult)
-        assert result.success is True
-        assert result.execution_mode == "visual"
-        assert result.action_taken is not None
-        assert result.action_taken.action_type == "navigate"
-        
-        # Verify agent calls
-        test_runner_agent.action_agent.execute_action.assert_called_once()
-        # Evaluation is now handled by Action Agent's AI analysis
-    
-    @pytest.mark.asyncio
-    async def test_execute_step_scripted_mode(
-        self, test_runner_agent, sample_test_plan
-    ):
-        """Test executing a step in scripted mode."""
-        step = sample_test_plan.steps[0]
-        # Set current test plan
-        test_runner_agent._current_test_plan = sample_test_plan
-        
-        # Add a scripted action
-        step_key = f"{sample_test_plan.plan_id}:{step.step_number}"
-        test_runner_agent._scripted_actions[step_key] = {
-            "action_type": "click",
-            "x": 960,
-            "y": 540,
-            "grid_cell": "M23",
-            "offset_x": 0.5,
-            "offset_y": 0.5
-        }
-        
-        # Execute
-        result = await test_runner_agent._execute_step(step, ExecutionMode.SCRIPTED)
-        
-        # Verify
-        assert result.success is True
-        assert result.execution_mode == "scripted"
-        assert result.action_taken.confidence == 1.0
-        
-        # Verify browser click was called
-        test_runner_agent.browser_driver.click.assert_called_once()
-        # Coordinates should be calculated from grid, not hardcoded
-        call_args = test_runner_agent.browser_driver.click.call_args[0]
-        assert len(call_args) == 2  # x, y coordinates
-    
-    @pytest.mark.asyncio
-    async def test_execute_step_hybrid_mode_fallback(
-        self, test_runner_agent, sample_test_plan
-    ):
-        """Test hybrid mode falling back to visual when scripted fails."""
-        step = sample_test_plan.steps[0]
-        # Set current test plan
-        test_runner_agent._current_test_plan = sample_test_plan
-        
-        # Add a scripted action that will fail
-        step_key = f"{sample_test_plan.plan_id}:{step.step_number}"
-        test_runner_agent._scripted_actions[step_key] = {
-            "action_type": "click",
-            "grid_cell": "M23",
-            "offset_x": 0.5,
-            "offset_y": 0.5
-        }
-        
-        # Make scripted execution fail on first call, succeed on second
-        test_runner_agent.browser_driver.click.side_effect = [
-            Exception("Element not found"),  # First call fails (scripted)
-            None  # Second call succeeds (visual fallback)
-        ]
-        
-        # Execute
-        result = await test_runner_agent._execute_step(step, ExecutionMode.HYBRID)
-        
-        # Verify fallback to visual
-        assert result.success is True
-        assert result.action_taken is not None
-        assert result.action_taken.action_type == "navigate"
-        test_runner_agent.action_agent.execute_action.assert_called_once()
-        # Browser click should be called once (scripted failed, visual succeeded through Action Agent)
-        assert test_runner_agent.browser_driver.click.call_count == 1
-    
-    def test_check_dependencies_met(self, test_runner_agent, sample_test_plan):
-        """Test dependency checking when dependencies are met."""
-        # Initialize test state with completed steps
-        test_runner_agent._test_state = TestState(
-            test_plan=sample_test_plan,
-            current_step=None,
-            completed_steps=[sample_test_plan.steps[0].step_id, sample_test_plan.steps[1].step_id],
-            failed_steps=[],
-            skipped_steps=[],
-            status=TestStatus.IN_PROGRESS,
-            start_time=datetime.now(timezone.utc),
-            error_count=0,
-            warning_count=0
+        """Test test plan execution with a critical failure."""
+        # Make the login button click fail
+        from src.core.enhanced_types import (
+            AIAnalysis,
+            ValidationResult,
+            CoordinateResult,
+            ExecutionResult,
+            EnhancedActionResult
         )
         
-        # Check step 4 which depends on steps 2 and 3
-        step4 = sample_test_plan.steps[3]
-        assert test_runner_agent._check_dependencies(step4) is False  # Step 3 not completed
-        
-        # Add step 3
-        test_runner_agent._test_state.completed_steps.append(sample_test_plan.steps[2].step_id)
-        
-        # Now dependencies should be met
-        assert test_runner_agent._check_dependencies(step4) is True
-    
-    def test_check_dependencies_failed(self, test_runner_agent, sample_test_plan):
-        """Test dependency checking when a dependency failed."""
-        # Initialize test state with failed step
-        test_runner_agent._test_state = TestState(
-            test_plan=sample_test_plan,
-            current_step=None,
-            completed_steps=[],  # Step 1 not in completed steps
-            failed_steps=[sample_test_plan.steps[0].step_id],  # Step 1 failed
-            skipped_steps=[],
-            status=TestStatus.IN_PROGRESS,
-            start_time=datetime.now(timezone.utc),
-            error_count=1,
-            warning_count=0
-        )
-        
-        # Check step 2 which depends on step 1
-        step2 = sample_test_plan.steps[1]
-        assert test_runner_agent._check_dependencies(step2) is False
-    
-    def test_record_action(self, test_runner_agent, sample_test_plan):
-        """Test recording successful actions."""
-        test_runner_agent._current_test_plan = sample_test_plan
-        step = sample_test_plan.steps[0]
-        
-        action = ActionResult(
-            action_type="click",
-            grid_cell="B7",
-            offset_x=0.3,
-            offset_y=0.7,
-            confidence=0.95,
-            requires_refinement=False
-        )
-        
-        # Record action
-        test_runner_agent._record_action(step, action)
-        
-        # Verify
-        step_key = f"{sample_test_plan.plan_id}:{step.step_number}"
-        assert step_key in test_runner_agent._scripted_actions
-        recorded = test_runner_agent._scripted_actions[step_key]
-        assert recorded["action_type"] == "click"
-        assert recorded["grid_cell"] == "B7"
-        assert recorded["offset_x"] == 0.3
-        assert recorded["offset_y"] == 0.7
-        # No pixel coordinates should be stored anymore
-    
-    @pytest.mark.asyncio
-    async def test_determine_execution_mode(
-        self, test_runner_agent, sample_test_plan
-    ):
-        """Test execution mode determination."""
-        test_runner_agent._current_test_plan = sample_test_plan
-        step = sample_test_plan.steps[0]
-        
-        # No scripted action - should use visual
-        mode = await test_runner_agent._determine_execution_mode(step)
-        assert mode == ExecutionMode.VISUAL
-        
-        # Add scripted action
-        step_key = f"{sample_test_plan.plan_id}:{step.step_number}"
-        test_runner_agent._scripted_actions[step_key] = {"action_type": "click"}
-        
-        # Should now use hybrid
-        mode = await test_runner_agent._determine_execution_mode(step)
-        assert mode == ExecutionMode.HYBRID
-    
-    @pytest.mark.asyncio
-    async def test_get_next_action(self, test_runner_agent, sample_test_plan):
-        """Test getting next action recommendation."""
-        test_runner_agent.call_openai = AsyncMock(return_value={
-            "content": "Proceed with the current step as planned"
-        })
-        
-        test_state = TestState(
-            test_plan=sample_test_plan,
-            current_step=sample_test_plan.steps[0],
-            completed_steps=[],
-            failed_steps=[],
-            skipped_steps=[],
-            status=TestStatus.IN_PROGRESS,
-            start_time=datetime.now(timezone.utc),
-            error_count=0,
-            warning_count=0
-        )
-        
-        # Get next action
-        recommendation = await test_runner_agent.get_next_action(
-            sample_test_plan, test_state
-        )
-        
-        assert recommendation == "Proceed with the current step as planned"
-        test_runner_agent.call_openai.assert_called_once()
-    
-    @pytest.mark.asyncio
-    async def test_analyze_progress(self, test_runner_agent, sample_test_plan):
-        """Test AI-based progress analysis."""
-        test_runner_agent._current_test_plan = sample_test_plan
-        test_runner_agent._test_state = TestState(
-            test_plan=sample_test_plan,
-            current_step=sample_test_plan.steps[2],
-            completed_steps=[sample_test_plan.steps[0].step_id, sample_test_plan.steps[1].step_id],
-            failed_steps=[],
-            skipped_steps=[],
-            status=TestStatus.IN_PROGRESS,
-            start_time=datetime.now(timezone.utc),
-            error_count=0,
-            warning_count=0
-        )
-        
-        test_runner_agent._execution_history = [
-            TestStepResult(
-                step=sample_test_plan.steps[0],
-                success=True,
-                action_taken=None,
-                actual_result="Success",
-                execution_mode="visual"
-            ),
-            TestStepResult(
-                step=sample_test_plan.steps[1],
-                success=True,
-                action_taken=None,
-                actual_result="Success",  
-                execution_mode="visual"
-            ),
-        ]
-        
-        test_runner_agent.call_openai = AsyncMock(return_value={
-            "content": json.dumps({
-                "assessment": "Test progressing smoothly",
-                "concerns": [],
-                "recommendations": ["Continue with login flow"]
-            })
-        })
-        
-        # Analyze progress
-        await test_runner_agent._analyze_progress()
-        
-        # Verify AI was called and analysis stored
-        test_runner_agent.call_openai.assert_called_once()
-        assert "latest_analysis" in test_runner_agent._test_state.context
-        analysis = test_runner_agent._test_state.context["latest_analysis"]
-        assert analysis["assessment"] == "Test progressing smoothly"
-    
-    @pytest.mark.asyncio
-    async def test_mark_step_skipped(self, test_runner_agent, sample_test_plan):
-        """Test marking a step as skipped."""
-        test_runner_agent._current_test_plan = sample_test_plan
-        test_runner_agent._test_state = TestState(
-            test_plan=sample_test_plan,
-            current_step=sample_test_plan.steps[3],
-            completed_steps=[sample_test_plan.steps[0].step_id, sample_test_plan.steps[1].step_id, sample_test_plan.steps[2].step_id],
-            failed_steps=[],
-            skipped_steps=[],
-            status=TestStatus.IN_PROGRESS,
-            start_time=datetime.now(timezone.utc),
-            error_count=0,
-            warning_count=0
-        )
-        test_runner_agent._execution_history = []
-        test_runner_agent._step_index = 3  # Current step index
-        
-        # Mark step as skipped
-        step = sample_test_plan.steps[3]
-        test_runner_agent._mark_step_skipped(step)
-        
-        # Verify
-        assert len(test_runner_agent._execution_history) == 1
-        result = test_runner_agent._execution_history[0]
-        assert result.success is False
-        assert "Skipped due to unmet dependencies" in result.actual_result
-        assert result.execution_mode == "skipped"
-        assert step.step_id in test_runner_agent._test_state.skipped_steps
-    
-    @pytest.mark.asyncio
-    async def test_judge_final_test_result_success(self, test_runner_agent, sample_test_plan):
-        """Test final judgment on successful test execution."""
-        # Set up completed test state
-        test_runner_agent._current_test_plan = sample_test_plan
-        test_runner_agent._test_state = TestState(
-            test_plan=sample_test_plan,
-            current_step=None,
-            completed_steps=[s.step_id for s in sample_test_plan.steps],
-            failed_steps=[],
-            skipped_steps=[],
-            status=TestStatus.COMPLETED,
-            start_time=datetime.now(timezone.utc),
-            end_time=datetime.now(timezone.utc),
-            error_count=0,
-            warning_count=0
-        )
-        
-        # Add successful execution history
-        test_runner_agent._execution_history = [
-            TestStepResult(
-                step=step,
-                success=True,
-                action_taken=ActionResult(
-                    action_type=step.action_instruction.action_type.value,
-                    grid_cell="M23",
-                    confidence=0.95
+        async def mock_execute_with_failure(test_step, test_context, screenshot=None):
+            # Fail on step 4 of login test case (click login button)
+            if "login button" in test_step.action:
+                return EnhancedActionResult(
+                    test_step_id=test_step.step_id,
+                    test_step=test_step,
+                    test_context=test_context,
+                    validation=ValidationResult(
+                        valid=True,
+                        reasoning="Action is valid",
+                        confidence=0.95,
+                        concerns=[],
+                        suggestions=[]
+                    ),
+                    coordinates=CoordinateResult(
+                        grid_cell="M23",
+                        grid_coordinates=(960, 540),
+                        offset_x=0.5,
+                        offset_y=0.5,
+                        confidence=0.95,
+                        reasoning="Element found",
+                        refined=False,
+                        refinement_details=None
+                    ),
+                    execution=ExecutionResult(
+                        success=False,
+                        execution_time_ms=1000.0,
+                        error_message="Element not found: login button",
+                        error_traceback=None,
+                        browser_logs=[],
+                        network_activity=[]
+                    ),
+                    ai_analysis=AIAnalysis(
+                        success=False,
+                        confidence=0.9,
+                        actual_outcome="Login button not found on page",
+                        matches_expected=False,
+                        ui_changes=[],
+                        recommendations=["Check if page loaded correctly"],
+                        anomalies=["Login button missing"]
+                    ),
+                    overall_success=False,
+                    failure_phase="execution"
+                )
+            # All other actions succeed
+            return EnhancedActionResult(
+                test_step_id=test_step.step_id,
+                test_step=test_step,
+                test_context=test_context,
+                validation=ValidationResult(
+                    valid=True,
+                    reasoning="Valid",
+                    confidence=0.95,
+                    concerns=[],
+                    suggestions=[]
                 ),
-                actual_result="Step completed successfully",
-                execution_mode="visual",
-                action_result_details={
-                    "validation_passed": True,
-                    "execution_success": True,
-                    "ai_analysis": {"success": True}
-                }
+                coordinates=CoordinateResult(
+                    grid_cell="M23",
+                    grid_coordinates=(960, 540),
+                    offset_x=0.5,
+                    offset_y=0.5,
+                    confidence=0.95,
+                    reasoning="Element found",
+                    refined=False,
+                    refinement_details=None
+                ),
+                execution=ExecutionResult(
+                    success=True,
+                    execution_time_ms=500.0,
+                    error_message=None,
+                    error_traceback=None,
+                    browser_logs=[],
+                    network_activity=[]
+                ),
+                ai_analysis=AIAnalysis(
+                    success=True,
+                    confidence=0.9,
+                    actual_outcome="Success",
+                    matches_expected=True,
+                    ui_changes=[],
+                    recommendations=[],
+                    anomalies=[]
+                ),
+                overall_success=True
             )
-            for step in sample_test_plan.steps
+        
+        mock_action_agent.execute_action = AsyncMock(side_effect=mock_execute_with_failure)
+        
+        # Mock AI responses
+        test_runner.call_openai = AsyncMock()
+        test_runner.call_openai.side_effect = [
+            # Step interpretations
+            {"content": json.dumps({"actions": [{"type": "navigate", "target": "https://example.com/login", "critical": True}]})},
+            {"content": json.dumps({"actions": [{"type": "type", "target": "username field", "value": "testuser", "critical": True}]})},
+            {"content": json.dumps({"actions": [{"type": "type", "target": "password field", "value": "testpass", "critical": True}]})},
+            {"content": json.dumps({"actions": [{"type": "click", "target": "login button", "critical": True}]})},
+            # Outcome verifications (first 3 succeed)
+            {"content": json.dumps({"success": True, "actual_outcome": "Action completed", "confidence": 0.95})},
+            {"content": json.dumps({"success": True, "actual_outcome": "Action completed", "confidence": 0.95})},
+            {"content": json.dumps({"success": True, "actual_outcome": "Action completed", "confidence": 0.95})},
+            # Blocker analysis for failed step
+            {"content": json.dumps({"is_blocker": True, "reasoning": "Login is critical"})},
+            # Second test case is blocked
         ]
         
-        # Mock AI judgment
-        test_runner_agent.call_openai = AsyncMock(return_value={
-            "content": json.dumps({
-                "overall_success": True,
-                "confidence": 0.95,
-                "reasoning": "All test steps completed successfully",
-                "key_issues": [],
-                "recommendations": []
-            })
-        })
+        # Execute test plan
+        report = await test_runner.execute_test_plan(sample_hierarchical_test_plan)
         
-        # Get final judgment
-        judgment = await test_runner_agent.judge_final_test_result()
+        # Verify failure handling
+        assert report.status == TestStatus.FAILED
+        assert len(report.test_cases) == 2
         
-        # Verify
-        assert judgment["overall_success"] is True
-        assert judgment["confidence"] == 0.95
-        assert judgment["execution_stats"]["success_rate"] == 1.0
-        assert len(judgment["key_issues"]) == 0
+        # First test case should fail
+        login_result = report.test_cases[0]
+        assert login_result.status == TestStatus.FAILED
+        assert login_result.steps_completed == 3  # First 3 steps succeeded
+        assert login_result.steps_failed == 1  # Login button click failed
+        
+        # Second test case should be blocked (prerequisites not met)
+        logout_result = report.test_cases[1]
+        assert logout_result.status == TestStatus.BLOCKED
+        
+        # Verify bug was reported
+        assert len(report.bugs) == 1
+        bug = report.bugs[0]
+        assert bug.severity == BugSeverity.HIGH
+        assert bug.error_type == "element_not_found"
+        assert "login button" in bug.description.lower()
+        assert bug.step_number == 4
+        
+        # Verify summary reflects failures
+        assert report.summary.failed_test_cases == 1
+        assert report.summary.success_rate < 1.0
     
     @pytest.mark.asyncio
-    async def test_judge_final_test_result_with_failures(self, test_runner_agent, sample_test_plan):
-        """Test final judgment on test with failures."""
-        # Set up test state with failures
-        test_runner_agent._current_test_plan = sample_test_plan
-        test_runner_agent._test_state = TestState(
-            test_plan=sample_test_plan,
-            current_step=None,
-            completed_steps=[sample_test_plan.steps[0].step_id],
-            failed_steps=[sample_test_plan.steps[1].step_id],
-            skipped_steps=[sample_test_plan.steps[2].step_id, sample_test_plan.steps[3].step_id],
-            status=TestStatus.FAILED,
-            start_time=datetime.now(timezone.utc),
-            end_time=datetime.now(timezone.utc),
-            error_count=1,
-            warning_count=0
+    async def test_intelligent_step_interpretation(self, test_runner):
+        """Test that steps are intelligently decomposed into multiple actions."""
+        # Create a complex step
+        complex_step = TestStep(
+            step_number=1,
+            description="Add product to cart",
+            action="Search for 'laptop', select first result, and add to cart",
+            expected_result="Product is added to cart and cart count shows 1"
         )
         
-        # Add mixed execution history
-        test_runner_agent._execution_history = [
-            TestStepResult(
-                step=sample_test_plan.steps[0],
-                success=True,
-                action_taken=None,
-                actual_result="Success",
-                execution_mode="visual"
-            ),
-            TestStepResult(
-                step=sample_test_plan.steps[1],
-                success=False,
-                action_taken=None,
-                actual_result="Failed to find element",
-                execution_mode="visual",
-                action_result_details={
-                    "validation_passed": False,
-                    "validation_reasoning": "Element not visible",
-                    "execution_error": "Element not found"
-                }
-            )
-        ]
+        test_case = TestCase(
+            test_id="TC001",
+            name="Shopping Test",
+            description="Test shopping flow",
+            priority=TestCasePriority.HIGH,
+            steps=[complex_step]
+        )
         
-        # Mock AI judgment
-        test_runner_agent.call_openai = AsyncMock(return_value={
+        # Mock AI to decompose into multiple actions
+        test_runner.call_openai = AsyncMock(return_value={
             "content": json.dumps({
-                "overall_success": False,
-                "confidence": 0.85,
-                "reasoning": "Critical step failed, preventing test completion",
-                "key_issues": ["Username field not found", "Page may not have loaded"],
-                "recommendations": ["Verify page load", "Check element selectors"]
+                "actions": [
+                    {
+                        "type": "type",
+                        "target": "search box",
+                        "value": "laptop",
+                        "description": "Type search query",
+                        "critical": True
+                    },
+                    {
+                        "type": "click",
+                        "target": "search button",
+                        "description": "Submit search",
+                        "critical": True
+                    },
+                    {
+                        "type": "wait",
+                        "target": "search results",
+                        "description": "Wait for results to load",
+                        "critical": False
+                    },
+                    {
+                        "type": "click",
+                        "target": "first product",
+                        "description": "Click first search result",
+                        "critical": True
+                    },
+                    {
+                        "type": "scroll",
+                        "target": "add to cart button",
+                        "description": "Scroll to add to cart button",
+                        "critical": False
+                    },
+                    {
+                        "type": "click",
+                        "target": "add to cart button",
+                        "description": "Add product to cart",
+                        "critical": True
+                    },
+                    {
+                        "type": "assert",
+                        "target": "cart count",
+                        "value": "1",
+                        "description": "Verify cart count",
+                        "critical": True
+                    }
+                ]
             })
         })
         
-        # Get final judgment
-        judgment = await test_runner_agent.judge_final_test_result()
+        # Execute step interpretation
+        actions = await test_runner._interpret_step(
+            complex_step, test_case, TestCaseResult(
+                case_id=test_case.case_id,
+                test_id=test_case.test_id,
+                name=test_case.name,
+                status=TestStatus.IN_PROGRESS,
+                started_at=datetime.now(timezone.utc),
+                steps_total=1,
+                steps_completed=0,
+                steps_failed=0
+            )
+        )
         
-        # Verify
-        assert judgment["overall_success"] is False
-        assert judgment["confidence"] == 0.85
-        assert judgment["execution_stats"]["success_rate"] == 0.25  # 1/4 steps
-        assert len(judgment["key_issues"]) == 2
-        assert len(judgment["recommendations"]) == 2
+        # Verify decomposition
+        assert len(actions) == 7
+        assert actions[0]["type"] == "type"
+        assert actions[0]["value"] == "laptop"
+        assert actions[6]["type"] == "assert"
+        assert actions[6]["value"] == "1"
+        
+        # Verify critical vs non-critical actions
+        assert actions[2]["critical"] is False  # Wait is non-critical
+        assert actions[4]["critical"] is False  # Scroll is non-critical
+        assert actions[5]["critical"] is True   # Add to cart is critical
+    
+    @pytest.mark.asyncio
+    async def test_bug_report_generation(self, test_runner):
+        """Test detailed bug report generation for failures."""
+        # Create a failed step result
+        step = TestStep(
+            step_number=2,
+            description="Submit form",
+            action="Click submit button",
+            expected_result="Form is submitted successfully"
+        )
+        
+        test_case = TestCase(
+            case_id=uuid4(),
+            test_id="TC001",
+            name="Form Submission Test",
+            description="Test form submission",
+            priority=TestCasePriority.HIGH,
+            steps=[step]
+        )
+        
+        step_result = StepResult(
+            step_id=step.step_id,
+            step_number=2,
+            status=TestStatus.FAILED,
+            started_at=datetime.now(timezone.utc),
+            completed_at=datetime.now(timezone.utc),
+            action="Click submit button",
+            expected_result="Form is submitted successfully",
+            actual_result="Submit button not found",
+            error_message="Element not found: submit button",
+            screenshot_after="/screenshots/error.png"
+        )
+        
+        # Set up test runner context
+        test_runner._current_test_plan = TestPlan(
+            plan_id=uuid4(),
+            name="Test Plan",
+            description="Test",
+            requirements_source="PRD",
+            test_cases=[test_case]
+        )
+        
+        # Create a test case result for the function
+        case_result = TestCaseResult(
+            case_id=test_case.case_id,
+            test_id=test_case.test_id,
+            name=test_case.name,
+            status=TestStatus.IN_PROGRESS,
+            started_at=datetime.now(timezone.utc),
+            steps_total=1,
+            steps_completed=0,
+            steps_failed=1,
+            step_results=[step_result]  # Include the failed step
+        )
+        
+        # Generate bug report
+        bug_report = await test_runner._create_bug_report(
+            step_result, step, test_case, case_result
+        )
+        
+        # Verify bug report
+        assert bug_report is not None
+        assert bug_report.severity == BugSeverity.HIGH
+        assert bug_report.error_type == "element_not_found"
+        assert bug_report.step_number == 2
+        assert "submit button" in bug_report.description
+        assert bug_report.expected_result == "Form is submitted successfully"
+        assert bug_report.actual_result == "Submit button not found"
+        assert bug_report.screenshot_path == "/screenshots/error.png"
+        assert len(bug_report.reproduction_steps) > 0
+    
+    @pytest.mark.asyncio
+    async def test_living_document_updates(self, test_runner, tmp_path):
+        """Test that reports are saved periodically as living documents."""
+        # Create a simple test plan
+        test_case = TestCase(
+            test_id="TC001",
+            name="Test Case",
+            description="Test",
+            priority=TestCasePriority.MEDIUM,
+            steps=[
+                TestStep(
+                    step_number=1,
+                    description="Step 1",
+                    action="Do something",
+                    expected_result="Success"
+                )
+            ]
+        )
+        
+        test_plan = TestPlan(
+            name="Test Plan",
+            description="Test",
+            requirements_source="PRD",
+            test_cases=[test_case]
+        )
+        
+        # Mock minimal AI responses
+        test_runner.call_openai = AsyncMock()
+        test_runner.call_openai.return_value = {
+            "content": json.dumps({
+                "actions": [{"type": "click", "target": "button", "critical": True}]
+            })
+        }
+        
+        # Start execution
+        test_runner._test_report = TestReport(
+            test_plan_id=test_plan.plan_id,
+            test_plan_name=test_plan.name,
+            started_at=datetime.now(timezone.utc),
+            status=TestStatus.IN_PROGRESS
+        )
+        
+        # Save report multiple times (simulating periodic saves)
+        await test_runner._save_report()
+        
+        # Verify files exist
+        report_dir = Path(test_runner.report_dir)
+        json_files = list(report_dir.glob("*.json"))
+        md_files = list(report_dir.glob("*.md"))
+        
+        assert len(json_files) == 1
+        assert len(md_files) == 1
+        
+        # Read and verify JSON content
+        with open(json_files[0], "r") as f:
+            saved_report = json.load(f)
+        
+        assert saved_report["test_plan_name"] == "Test Plan"
+        assert saved_report["status"] == "in_progress"
+        
+        # Update report and save again
+        test_runner._test_report.status = TestStatus.COMPLETED
+        await test_runner._save_report()
+        
+        # Verify updated content
+        with open(json_files[0], "r") as f:
+            updated_report = json.load(f)
+        
+        assert updated_report["status"] == "completed"
+    
+    @pytest.mark.asyncio
+    async def test_failure_classification(self, test_runner):
+        """Test that failures are correctly classified as blockers or non-blockers."""
+        # Test navigation error (should be blocker)
+        nav_result = StepResult(
+            step_id=uuid4(),
+            step_number=1,
+            status=TestStatus.FAILED,
+            started_at=datetime.now(timezone.utc),
+            completed_at=datetime.now(timezone.utc),
+            action="Navigate to page",
+            expected_result="Page loads",
+            actual_result="Navigation failed",
+            error_message="Navigation timeout"
+        )
+        
+        is_blocker = await test_runner._is_blocker_failure(nav_result)
+        assert is_blocker is True
+        
+        # Test element not found (should check with AI)
+        test_runner.call_openai = AsyncMock(return_value={
+            "content": json.dumps({
+                "is_blocker": True,
+                "reasoning": "Cannot proceed without this element"
+            })
+        })
+        
+        element_result = StepResult(
+            step_id=uuid4(),
+            step_number=2,
+            status=TestStatus.FAILED,
+            started_at=datetime.now(timezone.utc),
+            completed_at=datetime.now(timezone.utc),
+            action="Click button",
+            expected_result="Button clicked",
+            actual_result="Button not found",
+            error_message="Element not found"
+        )
+        
+        is_blocker = await test_runner._is_blocker_failure(element_result)
+        assert is_blocker is True
+        test_runner.call_openai.assert_called_once()
+    
+    @pytest.mark.asyncio
+    async def test_summary_calculation(self, test_runner):
+        """Test summary statistics calculation."""
+        # Create test report with mixed results
+        test_runner._test_report = TestReport(
+            test_plan_id=uuid4(),
+            test_plan_name="Test Plan",
+            started_at=datetime.now(timezone.utc),
+            completed_at=datetime.now(timezone.utc),
+            status=TestStatus.COMPLETED,
+            test_cases=[
+                TestCaseResult(
+                    case_id=uuid4(),
+                    test_id="TC001",
+                    name="Test 1",
+                    status=TestStatus.COMPLETED,
+                    started_at=datetime.now(timezone.utc),
+                    completed_at=datetime.now(timezone.utc),
+                    steps_total=3,
+                    steps_completed=3,
+                    steps_failed=0
+                ),
+                TestCaseResult(
+                    case_id=uuid4(),
+                    test_id="TC002",
+                    name="Test 2",
+                    status=TestStatus.FAILED,
+                    started_at=datetime.now(timezone.utc),
+                    completed_at=datetime.now(timezone.utc),
+                    steps_total=2,
+                    steps_completed=1,
+                    steps_failed=1
+                )
+            ],
+            bugs=[
+                BugReport(
+                    step_id=uuid4(),
+                    test_case_id=uuid4(),
+                    test_plan_id=uuid4(),
+                    step_number=1,
+                    description="Bug 1",
+                    severity=BugSeverity.CRITICAL,
+                    error_type="error",
+                    expected_result="Success",
+                    actual_result="Failure"
+                ),
+                BugReport(
+                    step_id=uuid4(),
+                    test_case_id=uuid4(),
+                    test_plan_id=uuid4(),
+                    step_number=2,
+                    description="Bug 2",
+                    severity=BugSeverity.MEDIUM,
+                    error_type="error",
+                    expected_result="Success",
+                    actual_result="Failure"
+                )
+            ]
+        )
+        
+        # Calculate summary
+        summary = test_runner._calculate_summary()
+        
+        # Verify calculations
+        assert summary.total_test_cases == 2
+        assert summary.completed_test_cases == 1
+        assert summary.failed_test_cases == 1
+        assert summary.total_steps == 5
+        assert summary.completed_steps == 4
+        assert summary.failed_steps == 1
+        assert summary.success_rate == 0.8  # 4/5 steps
+        assert summary.critical_bugs == 1
+        assert summary.medium_bugs == 1
+        assert summary.high_bugs == 0
+        assert summary.low_bugs == 0
